fix: Member 순환참조
